Optimize is_prime
The first optimization takes advantage of the fact that two is the only even prime. Thus we can check if a number is even and as long as its greater than 2, we know that it is not prime.

Our second optimization takes advantage of the fact that when checking factors, we only need to check odd factors up to the square root of a number. Consider a number  𝑛  decomposed into factors  𝑛=𝑎𝑏 . There are two cases, either  𝑛  is prime and without loss of generality,  𝑎=𝑛,𝑏=1  or  𝑛  is not prime and  𝑎,𝑏≠𝑛,1 . In this case, if  𝑎>√n , then  𝑏<√n . So we only need to check all possible values of  𝑏  and we get the values of  𝑎  for free! This means that even the simple method of checking factors will increase in complexity as a square root compared to the size of the number instead of linearly.

We have to  write a function which will find all prime numbers up to and including 𝑛.