𝑘 -means clustering algorithm
The algorithm will require us to keep track of two clusters, each of which have a list of points and a center (which is another point, not necessarily one of the points we are clustering). After making an initial guess at the center of the two clusters,  𝐶1  and  𝐶2 , the steps proceed as follows.
1. Assign each point to  𝐶1  or  𝐶2  based on whether the point is closer to the center of  𝐶1  or  𝐶2 .
2. Recalculate the center of  𝐶1  and  𝐶2  based on the contained points.
This algorithm will terminate in general when the assignments no longer change. For this question, we would like you to initialize one cluster at (1, 0) and the other at (-1, 0).
The returned values should be the two centers of the clusters ordered by greatest x value. Please return these as a list of numeric tuples  [(𝑥1,𝑦1),(𝑥2,𝑦2)] 
In order to accomplish this we will create a class called cluster which has two methods besides __init__ which you will need to write. The first method update will update the center of the Cluster given the points contained in the attribute points. Remember, you after updating the center of the cluster, you will want to reassign the points and thus remove previous assignments. The other method add_point will add a point to the points attribute.